Avoid a divide-by-zero in the currentTime algorithm
Previously if startScrollOffset == endScrollOffset then the final step
of the currentTime algorithm became:

(current scroll offset / 0) * effective time range

Since division by zero is undefined, insert a step to bail if this is
going to happen.

Note that since startScrollOffset and endScrollOffset can be calc()
values, we don't know until resolution whether they are equal, and they
may be equal in one call to currentTime and then not equal later.

Fixes #21